done with back end

+++ project1/app/Http/Controllers/AppointmentController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Models\Available_appointment;
 use App\Models\appointment;
 use Validator;
+use Illuminate\Support\Facades\Gate;
 
 
 class AppointmentController extends Controller
@@ -14,6 +15,9 @@ class AppointmentController extends Controller
 
 	//create
     public function create_appointment(Request $request){
+        if(!Gate::allows('create_appointment')){
+            return "403";
+        }
     	$validator=Validator::make($request->all(),[
     		'date' => 'required|max:191|string',
     		'time' => 'required|max:191|string'
@@ -30,30 +34,37 @@ class AppointmentController extends Controller
 
     //update
     public function update_appointment(Request $request){
+        if(!Gate::allows('update_appointment')){
+            abort(403);
+        }
     	$validator=Validator::make($request->all(),[
     		'date' => 'required|max:191|string',
-    		'time' => 'required|max:191|string'
+    		'time' => 'required|max:191|string',
+            'updatedTime' => 'required|max:191|string'
     	]);
     	if ($validator->fails()) {
     		return $validator->errors();
     	}
-    	Available_appointment::where('date',$request->input('date'))->update(['time'=>$request->input('time')]);
+    	Available_appointment::where('date',$request->input('date'))->where('time',$request->input('time'))->update(['time'=>$request->input('updatedTime')]);
     	return "ok";   		
     }
 
     //delete
     public function delete_appointment(Request $request){
+        if(!Gate::allows('delete_appointment')){
+            abort(403);
+        }
     	$a=Available_appointment::where('date',$request->input('date'))->where('time',$request->input('time'));
     	$d=$a->delete();
     	$deleted="could not delete";
     	if($d==1){
-    		$deleted="deleted";
+    		$deleted="ok";
     	}
     	return $deleted;
     }
 
     //view
-    public function view_appointments(Request $request){
+    public function view_appointments(){
     	$a=Available_appointment::all();
     	foreach ($a as $b) {
     		return $b;
@@ -95,7 +106,7 @@ class AppointmentController extends Controller
         $d=$a->delete();
         $deleted="could not delete";
         if($d==1){
-            $deleted="deleted";
+            $deleted="ok";
         }
         return $deleted;
     }
@@ -112,6 +123,9 @@ class AppointmentController extends Controller
 
     //view all booked appointments
     public function view_all_booked_appointments(){
+        if(!Gate::allows('view_all_booked_appointments')){
+            abort(403);
+        }
         $a=appointment::all();
         foreach ($a as $b) {
             return $b;
