+++ project1/app/Http/Controllers/RecordsController.php
@@ -4,10 +4,26 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Record;
+use Illuminate\Support\Facades\Gate;
+use Validator;
 
 class RecordsController extends Controller
 {
     public function create_record(Request $request){
+    	if(!Gate::allows('create_record')){
+    		abort(403);
+    	}
+    	$validator=Validator::make($request->all(),[
+    		'name' => 'required|max:191|string',
+    		'address' => 'required|max:191|string',
+    		'phone' => 'required|max:191|string',
+    		'ill' => 'required|max:191|string',
+    		'status' => 'required|max:191|string',
+    		'user_id' => 'required|max:191|string'
+    	]);
+    	if ($validator->fails()) {
+    		return $validator->errors();
+    	}
     	Record::create([
             'name'=>$request->name,
             'address'=>$request->address,
@@ -16,22 +32,32 @@ class RecordsController extends Controller
             'status'=>$request->status,
             'user_id'=>$request->user_id,
         ]);
-
+        return "ok";
     }
     
     public function delete_record(Request $request){
+    	if(!Gate::allows('delete_record')){
+    		abort(403);
+    	}
     	$id=$request->id;
     	$record=Record::find($id);
     	$record->delete();
+    	return "ok";
     }
 
     public function view_record(Request $request){
+    	if(!Gate::allows('view_record')){
+    		abort(403);
+    	}
     	$id=$request->id;
     	$record=Record::find($id);
     	return $record;
     }
 
     public function view_all_records(){
+    	if(!Gate::allows('view_all_records')){
+    		abort(403);
+    	}
         $record=Record::all();
         foreach ($record as $r) {
             return $r;
@@ -44,18 +70,31 @@ class RecordsController extends Controller
         $a=Record::where('user_id',$id);
         foreach ($a as $b) {
             return $b;
-        }   
+        }      
     }
 
     public function search_record(Request $request){
+    	if(!Gate::allows('search_record')){
+    		abort(403);
+    	}
     	$name=$request->name;
     	$record=Record::where('name',$name);
-    	return $record;
+    	foreach ($record as $a) {
+            return $a;
+        }    
     }
 
     public function update_record(Request $request){
+    	if(!Gate::allows('update_record')){
+    		abort(403);
+    	}
+    	$validator=Validator::make($request->all(),[
+    		'status' => 'required|max:191|string'
+    	]);
+    	if ($validator->fails()) {
+    		return $validator->errors();
+    	}
     	$id=$request->id;
-    	$record=Record::where('name',$name);
     	Record::where('id',$id)->update(['status'=>$request->input('status')]);
     	return "ok"; 
     }
